VoiceTypingWindows/requirements.txt
faster-whisper==1.0.1
pystray==0.19.5
Pillow==10.2.0
pynput==1.7.6
pyautogui==0.9.54
pyperclip==1.8.2
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.4
pyinstaller==6.4.0
